+++ src/FluentDbTools/Example/Example.FluentDbTools.Migration/MigrationModels/AddPersonTable.cs
@@ -1,4 +1,5 @@
 using System;
+using System.IO;
 using Example.FluentDbTools.Database;
 using FluentDbTools.Common.Abstractions;
 using FluentDbTools.Migration.Abstractions;
@@ -30,10 +31,10 @@ namespace Example.FluentDbTools.Migration.MigrationModels
             // -
             // GlobalId is fetched from database:migration:schemaPrefix:tables:testing:GlobalId or database:schemaPrefix:tables:testing:GlobalId<br/>
             // ShortName is fetched from database:migration:schemaPrefix:tables:testing:ShortName or database:schemaPrefix:tables:testing:ShortName<br/>
-            var testChangeLogContext = new ChangeLogContext(this, Table.Testing) { EnabledTriggersAndViewsGeneration = TriggersAndViewsGeneration.Both};
+            var testChangeLogContext = new ChangeLogContext(this, Table.Testing) { EnabledTriggersAndViewsGeneration = TriggersAndViewsGeneration.Both };
 
             CreateParent();
-            
+
             // If "database:schemaPrefix:Id" or "database:migration:schemaPrefix:Id" have a value,
             // the tableName Person will be created as {SchemaPrefixId}Person.
             //
@@ -42,8 +43,9 @@ namespace Example.FluentDbTools.Migration.MigrationModels
             // the tableName Person will be created.
             var syntax = Create
                 .Table(Table.Person.GetPrefixedName(SchemaPrefixId)).InSchema(SchemaName)
+                .WithErrorFilter(1001)
                 .WithChangeLog(PersonLogContext)
-                .WithColumn(Table.Person + "Id").AsGuid().PrimaryKey().WithColumnDescription("Unique id.");
+                .WithColumn(Table.Person + "Id").AsGuid().WithColumnDescription("Unique id.");
 
             if (IsOracle())
             {
@@ -63,6 +65,7 @@ namespace Example.FluentDbTools.Migration.MigrationModels
 
                 .WithColumn("TestCol").AsGuid().WithColumnDescription("TestCol Guid")
                 .WithColumn(Column.Password).AsString().WithChangeLog(PersonLogContext).WithColumnDescription("password.")
+                .WithErrorFilter(1002)
                 .WithDefaultColumns() // Enable DefaultColumns functionality
                 .WithChangeLog(PersonLogContext)
                 .WithForeignKeyColumn("Parent".GetPrefixedName(SchemaPrefixId), this, "ParentId_FK", "ParentId").AsGuid().Nullable()
@@ -72,10 +75,12 @@ namespace Example.FluentDbTools.Migration.MigrationModels
                 .WithTableSequence(this);
 
             Alter.Table(Table.Person, this).WithDescription("description of " + PersonTableName).InSchema(SchemaName)
+                .WithErrorFilter(2001)
                 .WithChangeLog(PersonLogContext)
                 .AddColumn("NewCol").AsInt64().Nullable().WithColumnDescription("NewCol description")
                 .AlterColumn(Column.Password).AsFixedLengthString(255).Nullable()
                 .AlterColumn("TestCol").AsString(255).Nullable().WithColumnDescription("TestCol String(255)")
+                .WithErrorFilter(2002)
                 .WithChangeLog(PersonLogContext);
 
             Create.Table("Test", this)
@@ -99,11 +104,33 @@ namespace Example.FluentDbTools.Migration.MigrationModels
 
             Create.Column("CreateColumn")
                 .OnTable(Table.Person, this).AsInt16().Nullable()
+                .WithErrorFilter(3001)
                 .WithChangeLog(PersonLogContext);
 
             if (IsOracle())
             {
                 Execute.Sql($"create or replace synonym {SchemaName}.{Table.Testing}1 for {SchemaName}.{Table.Testing.GetPrefixedName(SchemaPrefixId)}");
+            }
+
+            if (IsOracle())
+            {
+                var sql =
+@"-- ErrorFilter = 6512, 955 	
+BEGIN 
+  EXECUTE IMMEDIATE 'CREATE TABLE {SchemaName}.{SchemaPrefixId}Person (PersonId RAW(16) DEFAULT sys_guid() NOT NULL, SequenceNumber NUMBER(10,0) NOT NULL, Alive NUMBER(1,0) NOT NULL, Username NVARCHAR2(255) NOT NULL, ExtraInformation BLOB, TestCol RAW(16) NOT NULL, Password NVARCHAR2(255) NOT NULL, ParentId_FK RAW(16))';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.PersonId IS ''Unique id.''';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.SequenceNumber IS ''sequence number.''';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.Alive IS ''Alive flag.''';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.Username IS ''username.''';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.ExtraInformation IS ''Extra Information as blob.''';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.TestCol IS ''TestCol Guid''';
+  EXECUTE IMMEDIATE 'COMMENT ON COLUMN {SchemaName}.{SchemaPrefixId}Person.Password IS ''password.'''; 
+END;";
+                sql = sql.ReplaceIgnoreCase("{SchemaName}", SchemaName).ReplaceIgnoreCase("{SchemaPrefixId}", SchemaPrefixId ?? string.Empty);
+                Execute.Sql(sql);
+                Execute.Script(Path.Combine("..","..", "..", "..","..", "Example","Example.FluentDbTools.Migration","MigrationModels","ExecuteSql.txt"));
+                
+                //Execute.EmbeddedScript("ExecuteEmbeddedSql.txt");
             }
         }
 
@@ -134,7 +161,7 @@ namespace Example.FluentDbTools.Migration.MigrationModels
         {
             var testChangeLogContext = new ChangeLogContext(this, Table.Testing);
             Delete.Table(GetPrefixedName(Table.Person))
-                .WithChangeLog(PersonLogContext,this)
+                .WithChangeLog(PersonLogContext, this)
                 .InSchema(SchemaName);
 
             Delete.Table(Table.Parent, this)