+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 <!-- the topmost header version must be set manually in the VERSION file -->
+### Version 1.2.35 2022-02-10
+- Added Errorfilter to Create and Alter syntax
+- Improved resolving migration log-file
+- Ignored Disposed exception for Migration RollbackTransaction 
+
 ### Version 1.2.34 2022-01-21
 - Improved IsInvalidDatabaseAdminException extension method.
 - Added tests for the IsInvalidDatabaseAdminException-method